Implement CR feedback
